+++ zallet-core/src/components/json_rpc/methods.rs
@@ -858,7 +858,14 @@ impl<C: Chain> RpcServer for RpcImpl<C> {
     }
 
     async fn view_transaction(&self, txid: &str) -> view_transaction::Response {
-        view_transaction::call(self.wallet().await?.as_ref(), self.chain().await?, txid).await
+        view_transaction::call(
+            self.wallet().await?.as_ref(),
+            #[cfg(zallet_build = "wallet")]
+            &self.keystore,
+            self.chain().await?,
+            txid,
+        )
+        .await
     }
 
     async fn stop(&self) -> stop::Response {